DEPLOYMENT/requirements.txt
streamlit
torch
torchvision
Pillow
fpdf
transformers
numpy
